to_datastore_requirements.txt
google-cloud-pubsub>=1.0.0
google-cloud-datastore>=1.1.0
